feat: Add user data to new posts in home.jsx
This commit adds functionality to fetch user data for new posts in the home.jsx file. When a new post is inserted, the user data is retrieved using the getUserData function from the userService module. The fetched user data is then added to the new post object before updating the posts state.